refactor: rename errorCorrectionLevel.ts to ErrorCorrectionLevel
