+++ example-app/SantokuApp/src/features/demo-qrcode/types/ErrorCorrectionLevel.ts
